+++ runtime/code_heap.h
@@ -3,4 +3,4 @@
 void rt_code_heap_create(void);
 void rt_code_heap_destroy(void);
 
-void* rt_code_heap_alloc(unsigned int size);
+void* rt_code_heap_alloc(size_t size);
