+++ runtime/support.h
@@ -9,11 +9,11 @@ typedef struct {
 	bool sealed;
 } rt_upval_t;
 
-typedef rt_value (*rt_compiled_closure_t)(rt_upval_t **upvals, rt_value obj, unsigned int argc, ...);
+typedef rt_value (*rt_compiled_closure_t)(rt_upval_t **upvals, rt_value obj, size_t argc, ...);
 
 void __stdcall rt_support_seal_upval(rt_upval_t *upval);
 
-rt_value rt_support_interpolate(unsigned int argc, ...);
+rt_value rt_support_interpolate(size_t argc, ...);
 
 rt_value __stdcall rt_support_get_const(rt_value obj, rt_value name);
 void __stdcall rt_support_set_const(rt_value obj, rt_value name, rt_value value);
